Moved tests to `tests` folders
Also renamed the files to match what they're testing, and simultaneously split `groups/tests.py` into fitting files.

Also, removed the empty `users/tests.py`.
